b43-fwcutter: unsupported file support
This patch makes fwcutter support only those files we know are working
and marks the other one unsupported. To allow developers to still work
with such files, an --unsupported command line option is added that
allows one to extract firmware from such unsupported files. Some code
restructuring was necessary to support the --unsupported flag with other
flags in any order.

Also, because this is a significant change, bump the version number.

Signed-off-by: Johannes Berg <johannes@sipsolutions.net>
Signed-off-by: Michael Buesch <mb@bu3sch.de>