ssb-sprom: Put SPROM data in a master table and add Rev. 8
Program ssb-sprom has gotten unwieldy as more versions of the SPROM have been
accommodated. As we now need to add Rev. 8 SPROMs for LP and N PHY devices, I
chose to rewrite the program now.

This patch converts the program to be driven by a table, which removes the need
for hard coding a number of revision-specific values. In fact, only two such
numbers are defined in the new header - the SPROM sizes for 128 and 440 byte
versions. Other changes include modifying the data order so that the variables
that are most likely to be changed are at the botton the list. I also removed
the SPROM revision from the list that could be modified. Changing this value
would BRICK the device. At the moment, only the first of the "per-path"
variables have been added for revisions 4, 5, and 8. As we gain more experience
with these variants, addition of new outputs may be necessary.

Signed-off-by: Larry Finger <Larry.Finger@lwfinger.net>
Signed-off-by: Michael Buesch <mb@bu3sch.de>