Changes in b43-fwcutter (V2)
The present version of b43-fwcutter extracts core 4 firmware from V4 drivers that
contain it even though b43 will not work with these devices. As this causes
confusion among the users, this patch suppresses such extraction.

Note: For b43legacy, firmware for core 5 devices is extracted. This capacity should
be retained as this overlap is used in the testing of b43legacy.

Signed-off-by: Larry Finger <Larry.Finger@lwfinger.net>
Signed-off-by: Michael Buesch <mb@bu3sch.de>