b43-fwcutter: Add support for generating a fw for brcmsmac
The normal firmware for brcmsmac is available in the linux-firmware git
repository, but this firmware does not work on a bcm4716 devices with a
ieee80211 core rev 16. When using a firmware extracted from the
proprietary Broadcom driver the device works with brcmsmac.

I tested a firmware version 666.2 extracted from the proprietary driver
with brcmsmac on a BCM4716 and a BCM43224 and haven't seen any
problems. The interface between the firmware and the driver changed
some times, brcmsmac only support firmware version 598 and newer.

For some entries in the firmware used by brcmsmac, fwcutter has to be
more extended, so we are adding dummy entires for them in
brcmsmac_add_dummy_entries(). Currently these entries are not used by
brcmsmac, but it is checked if they are in the archive.

I would still appreciate if Broadcom could release a firmware with
support for the ieee80211 core found in the bcm4716.

Signed-off-by: Hauke Mehrtens <hauke@hauke-m.de>
Signed-off-by: Michael Buesch <m@bues.ch>